[서버 :: 라우터 추가, url 경로 변경으로 인한 요청 url들 변경]
