+++ server/server.js
@@ -5,6 +5,8 @@ const cors = require("cors");
 const yaml = require("js-yaml");
 const fs = require("fs");
 const mongoose= require("mongoose");
+const ownerPage = require("./routers/OwnerPageRouter");
+const shoppingBasket = require("./routers/ShoppingBasketRouter");
 
 const app = express();
 const form_data = multer();
@@ -28,74 +30,8 @@ app.use(cors({
 }));
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
-
-
-app.get("/", (req, res) => {
-    return res.send("서버 구동중");
-});
-
-//DB 조회가 완료된 후 send 해야하므로 비동기 처리 필수
-//비동기 처리하지 않으면 빈배열만 보내게 됨
-app.get("/Stats/loadData", async (req, res) => {
-    let result = {};
-    let idx = 0;
-
-    const pizzaName = [];
-    await db.collection("pizza").find().sort({"name": 1}).forEach((r) => {
-        pizzaName.push(r.name);
-    });
-
-    const amounts = [];
-    await db.collection("orderHistory").find({"ownerNumber": "1"}).sort({"menu": 1}).forEach((r) => {
-        idx = amounts[idx] === undefined ? idx + 1 : idx;
-        amounts.splice(idx, 0, amounts[idx] ? amounts[idx] + 1 : 1);
-    });
-
-    idx = 0;
-    const sales = [];
-    const dayOfTheWeek = [];
-    await db.collection("orderHistory").find({"ownerNumber": "1"}).sort({"date": 1}).forEach((r) => {
-        if (idx !== 0) {
-            if (dayOfTheWeek[idx - 1] !== r.date) {
-                dayOfTheWeek[idx] = r.date;
-                sales[idx++] = Number(r.amount);
-            } else {
-                sales[idx - 1] += Number(r.amount);
-            }
-        } else {
-            dayOfTheWeek[idx] = r.date;
-            sales[idx++] = Number(r.amount);
-        }
-
-    });
-
-    result = {result, pizzaName: pizzaName, amounts: amounts, sales: sales, dayOfTheWeek: dayOfTheWeek};
-    return res.json(result);
-})
-
-app.post("/MenuReg", form_data.single("img"), (req, res) => {
-    db.collection("pizza").insertOne({...req.body, img: req.file}).then(() => {
-        console.log("데이터 삽입 성공");
-    });
-    return res.json(req.body);
-})
-
-/*app.post("/tempDummyInsert", form_data.any(), (req, res) => {
-    console.log(req.body);
-    db.collection("orderHistory").insertOne(req.body).then(() => {
-        console.log("데이터 삽입 성공");
-    });
-    return res.json(req.body);
-})*/
-
-app.get("/test", async (req, res) => {
-    let result = [];
-    await db.collection("pizza").find().forEach((r) => {
-        result.push({pizzaImg: r.img, pizzaName: r.name, pizzaPrice: r.priceL, pizzaOption: "테스트"});
-    })
-    //TODO 피자 컬렉션이 아니라 장바구니 컬렉션에서 조회하기
-    return res.json(result);
-})
+app.use("/ownerPage", ownerPage(form_data, db));
+app.use("/shoppingBasket", shoppingBasket(db));
 
 app.listen(port, hostname, () => {
     console.log(`Server running at http://${hostname}:${port}/`);
