+++ src/components/owner_page/OwnerPageMenuReg.js
@@ -72,7 +72,7 @@ function OwnerPageMenuReg() {
         pizza.append("topping3", topping3);
         pizza.append("img", img);
 
-        axios.post("http://localhost:4000/MenuReg", pizza).then(r => console.log(r));
+        axios.post("http://localhost:4000/ownerPage/menuReg", pizza).then(r => console.log(r));
     }
 
 /*    //테스트할 때 사용했던 더미 거래내역 넣는 함수 취소 버튼에 등록하여 사용했음
@@ -84,7 +84,7 @@ const tempDummy = () => {
         orderHistory.append("menu", name);
         orderHistory.append("amount", priceL);
         orderHistory.append("date", `${today.getFullYear()}/${today.getMonth() + 1}/${today.getDate()}`);
-        axios.post("http://localhost:4000/tempDummyInsert", orderHistory).then(r => console.log(r));
+        axios.post("http://localhost:4000/ownerPage/tempDummyInsert", orderHistory).then(r => console.log(r));
     }*/
 
     return (
