+++ src/index.js
@@ -15,7 +15,6 @@ import InfoModify from './components/info_modify/InfoModify';
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-    <Review />
     <BrowserRouter>
         <OwnerPage/>
     </BrowserRouter>
